Refactor: Move components to docs directory and update imports
- Deleted EditableTable, HeadBar, Navbar, Sidebar, ToolBar, create-user-dialog, and edit-user-dialog components from their original locations.
- Created new versions of EditableTable, HeadBar, Navbar, Sidebar, ToolBar, create-user-dialog, and edit-user-dialog in the docs directory.
- Updated import paths in Dashboard and admin-dashboard pages to reflect new component locations.
- Added JWT_SECRET and ACCEPT_ALL_ORIGINS to the server environment variables.
- Modified auth controller to handle cookie settings for SameSite attribute.
