+++ client/react-frontend/src/components/admin/create-user-dialog.tsx
@@ -17,7 +17,7 @@ import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { useToast } from "@/hooks/use-toast"
-import type { User, CreateUserData } from "../types/user"
+import type { User, CreateUserData } from "../../types/user"
 
 interface CreateUserDialogProps {
   open: boolean
