+++ client/react-frontend/src/components/admin/edit-user-dialog.tsx
@@ -17,7 +17,7 @@ import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { useToast } from "@/hooks/use-toast"
-import type { User, UpdateUserData } from "../types/user"
+import type { User, UpdateUserData } from "../../types/user"
 
 interface EditUserDialogProps {
   user: User | null
