+++ client/react-frontend/src/components/docs/Navbar.tsx
@@ -1,5 +1,5 @@
 import type { FC } from "react";
-import IconButton from "./IconButton";
+import IconButton from "../IconButton";
 import { FaHome, FaSitemap } from "react-icons/fa";
 import { useNavigate } from "react-router-dom";
