+++ server/node-server/.env.example
@@ -8,13 +8,14 @@ INIT_DB_ON_START=
 DATABASE_URL=
 NODE_ENV=
 
-SECRET_SESSION=
-
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
 GEMINI_API_KEY=
 
 CLIENT_HOST=
 ALLOWED_ORIGINS=
+
+JWT_SECRET=
+ACCEPT_ALL_ORIGINS=
